refactor: extract players router from main.py (23 endpoints)
Part of main.py simplification plan - extracting all player-related
endpoints into dedicated router module.

Migrated endpoints:
- Core CRUD: /players (POST, GET), /players/{id} (GET, PUT, DELETE)
- Search: /players/all, /players/name/{name}
- Statistics: /players/{id}/statistics, analytics, profile-with-stats
- Advanced: /players/{id}/advanced-metrics, trends, insights, skill-rating
- Availability: /players/me/availability, /players/{id}/availability, /players/availability/all
- Email prefs: /players/me/email-preferences, /players/{id}/email-preferences
